+++ backend/src/controllers/core/books_controller.js
@@ -49,74 +49,3 @@ export async function getBook(req, res) {
       .json({ message: error.message || "Internal Server Error" });
   }
 }
-
-// FOR ADMINS ONLY
-export async function addBook(req, res) {
-  try {
-    const {
-      title,
-      description,
-      authorCode,
-      publisher,
-      publicationDate,
-      categoryIds,
-      images,
-      price,
-      quantity,
-      pages,
-      language,
-    } = req.body;
-
-    const newBook = await bookService.createBookWithAssets({
-      title,
-      description,
-      authorCode,
-      publisher,
-      publicationDate,
-      categoryIds,
-      images,
-      price,
-      quantity,
-      pages,
-      language,
-    });
-
-    res.status(201).json(newBook);
-  } catch (error) {
-    console.error("Error adding book", error);
-    res
-      .status(error.status || 500)
-      .json({ message: error.message || "Internal Server Error" });
-  }
-}
-export async function updateBook(req, res) {
-  try {
-    const {
-      title,
-      description,
-      authorCode,
-      publisher,
-      publicationDate,
-      categoryIds,
-      images,
-      price,
-    } = req.body;
-    const bookCode = req.params.id;
-    const updatedBook = await bookService.updateBookService({
-      bookCode,
-      title,
-      description,
-      authorCode,
-      publisher,
-      publicationDate,
-      categoryIds,
-      images,
-      price,
-    });
-
-    res.status(200).json(updatedBook);
-  } catch (error) {
-    console.error("Error updating book", error);
-    res.status(500).json({ message: "Internal Server Error" });
-  }
-}
