+++ backend/src/services/core/book_services.js
@@ -85,93 +85,3 @@ export async function getBookService(id) {
     throw error;
   }
 }
-// FOR ADMINS ONLY
-export async function createBookWithAssets(book) {
-  try {
-    const {
-      title,
-      description,
-      authorCode,
-      publisher,
-      publicationDate,
-      categoryIds,
-      images,
-      price,
-      quantity,
-      pages,
-      language,
-    } = book;
-
-    // Resolve author via public code
-    const author = await Author.findOne({ authorCode: authorCode });
-    if (!author) {
-      const error = new Error("Author not found");
-      error.status = 404;
-      throw error;
-    }
-
-    const newBook = new Book({
-      title,
-      description,
-      author: author._id, // <-- ObjectId reference
-      publisher,
-      publicationDate,
-      categories: categoryIds,
-      price,
-      pages,
-      language,
-    });
-    const savedBook = await newBook.save();
-    if (images?.length) {
-      await mediaServices.attachImagesToBookService(savedBook._id, images);
-    }
-
-    await inventoryService.initializeInventory(savedBook._id, quantity);
-    return { book };
-  } catch (error) {
-    throw error;
-  }
-}
-export async function updateBookService(book) {
-  try {
-    const {
-      bookCode,
-      title,
-      description,
-      authorCode,
-      publisher,
-      publicationDate,
-      categoryIds,
-      images,
-      price,
-      pages,
-      language,
-    } = book;
-    const updateBook = await Book.findOneAndUpdate(
-      { bookCode: bookCode },
-      {
-        title,
-        description,
-        authorCode,
-        publisher,
-        publicationDate,
-        categoryIds,
-        price,
-        pages,
-        language,
-      },
-      { new: true },
-    );
-    if (images?.length) {
-      await mediaServices.attachImagesToBookService(updateBook._id, images);
-    }
-    if (!book) {
-      const error = new Error("Book not found");
-      error.status = 404;
-      throw error;
-    }
-    return { updateBook };
-  } catch (error) {
-    throw error;
-  }
-}
